\name{simuResult}
\docType{data}
\alias{simuResult}
\title{The results for the analysis of the simulated data.}
\description{
 The simulation and analyis are described in iClusterPlus/inst/unitTests/test_iClusterPlus.R. 
}
\usage{data(simuResult)}
\format{
  list
}
\value{
  A list of objects returned by the iClusterPlus function.
} 
\references{
  iClusterPlus/inst/unitTests/test_iClusterPlus.R	
}
\keyword{datasets}
