\name{variation.hg18.v10.nov.2010}
\docType{data}
\alias{variation.hg18.v10.nov.2010}
\title{Human genome variants of the NCBI 36 (hg18) assembly}
\description{
 Human genome variants of the NCBI 36 (hg18) assembly
}
\usage{data(variation.hg18.v10.nov.2010)}
\format{
  data frame
}
\value{
  \item{variation.hg18.v10.nov.2010}{Human genome variants of the NCBI 36 (hg18) assembly}
} 
\references{
http://projects.tcag.ca/variation/tableview.asp?table=DGV_Content_Summary.txt
}
\keyword{datasets}
